Move progress tracker completion from `StatesDiscoverCmd` to `CmdExecution`.
